Added a node-migrate.js to the bin directory.
This will replace the migrate.js and is using the commander.js API to tidy up the code.

Command will be extracted into the lib directory in future commits as they're not a concern of the console.

Changed the node-migrate in the bin section of package.json to point to node-migrate.js.